Refactor IComponents
