--- src/SkyWalking.NetworkProtocol.Trace/SkyWalking.NetworkProtocol.Trace.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\..\build\common.props" />
-  <PropertyGroup>
-    <Description>SkyWalking trace protocol.</Description>
-    <AssemblyTitle>SkyWalking.NetworkProtocol.Trace</AssemblyTitle>
-    <AssemblyName>SkyWalking.NetworkProtocol.Trace</AssemblyName>
-    <PackageId>SkyWalking.NetworkProtocol.Trace</PackageId>
-    <PackageTags>SkyWalking</PackageTags>
-    <PackageReleaseNotes>
-    </PackageReleaseNotes>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-</Project>
